requirements.txt
aioredis==1.3.1
annotated-types==0.5.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.1.0
click==8.1.7
confusable-homoglyphs==3.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-registration==3.4
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-stripe-subscription==1.2.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.8
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydotplus==2.0.2
PyJWT==2.7.0
pyparsing==3.1.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.7.1
tzlocal==5.2
urllib3==2.0.3
vonage==3.13.0
vonage-jwt==1.1.0
Werkzeug==2.0.2
wrapt==1.16.0
zope.interface==6.2
